Add files via upload
players all across the world have cherished the iconic video game Snake. The player controls a snake that crawls around the screen while attempting to consume food, avoiding obstacles or running into its own tail. It is a straightforward but addictive game. The snake lengthens each time it consumes food, and the player must control it as it moves through the game’s environment while attempting to consume as much food as they can without colliding. As the snake lengthens, the game’s difficulty increases because it gets more difficult to move around the screen. The Snake game is renowned for its straightforward gameplay features, which make it simple to learn but challenging to master. In this article, we are going to design the Snake game using Python‘s Tkinter library but before we start the game